Implement simple metadata example - per-node count of contained items.
Note that the structures in this commit are NOT the final structures, it is very likely that the simple 'size' field will be replaced with a map<String, Any> with a reserved key for the item count. Similarly the mdata::ft_size function will also probably be replaced with a function mdataForKey : String -> Any. A mdata_join : (String, Any, Any) -> Any callback will probably also be added.